overflow fix

+++ src/exponential_mechanism/mechanism.py
@@ -15,7 +15,8 @@ class ExponentialMechanism:
 
     def probabilities(self, output):
         exponent = (self.eps * self.scores(output)) / (2 * self.sensitivity)
-        probabilities = np.exp(exponent.astype('float128'))
+        exponent = exponent.astype('float128')
+        probabilities = np.exp(exponent)
         probabilities = probabilities / np.sum(probabilities)
         return probabilities
